Making sure the helpers are compliant
