+++ Microsoft.Toolkit.Uwp/Helpers/BluetoothLEHelper/GattProtocolErrorParser.cs
@@ -1,18 +1,26 @@
-﻿// <copyright file="GattProtocolErrorParser.cs" company="Microsoft Corporation">
-// Copyright (c) Microsoft Corporation.  All rights reserved.
-// </copyright>
-//----------------------------------------------------------------------------------------------
+﻿// ******************************************************************
+// Copyright (c) Microsoft. All rights reserved.
+// This code is licensed under the MIT License (MIT).
+// THE CODE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+// IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH
+// THE CODE OR THE USE OR OTHER DEALINGS IN THE CODE.
+// ******************************************************************
+
 using Windows.Devices.Bluetooth.GenericAttributeProfile;
 
 namespace Microsoft.Toolkit.Uwp
 {
     /// <summary>
-    /// Helper function when working with <see cref="GattProtocolError"/>
+    /// Helper function when working with <see cref="GattProtocolError" />
     /// </summary>
     public static class GattProtocolErrorParser
     {
         /// <summary>
-        /// Helper to convert an gatt error value into a string
+        /// Helper to convert an Gatt error value into a string
         /// </summary>
         /// <param name="errorValue"> the byte error value.</param>
         /// <returns>String representation of the error</returns>
