+++ Microsoft.Toolkit.Uwp/Microsoft.Toolkit.Uwp.csproj
@@ -50,6 +50,7 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Helpers\BluetoothLEHelper\GattNativeUuid.cs" />
     <Compile Include="Helpers\ClipboardHelper.cs" />
     <Compile Include="Helpers\BluetoothLEHelper\BluetoothLEHelper.cs" />
     <Compile Include="Helpers\BluetoothLEHelper\BytePadder.cs" />
